Updates to the Synapse R client:
added methods for maintaining object cache linked to objects held by a Layer entity, but not derived from the associated files. Updated addObject, deleteObject, etc. methods to automatically keep the object cache in sync with the object file cache.

TODO: still need to implement storeEntityObjects for this to be useful (PLFM-585).

git-svn-id: https://sagebionetworks.jira.com/svn/PLFM/trunk/client/rSynapseClient@4499 e6edf6fb-f266-4316-afb4-e53d95876a76